+++ src/main/java/com/cg/controller/admin/ExamPaperController.java
@@ -43,7 +43,7 @@ public class ExamPaperController {
      * @return
      */
     @RequestMapping(value = "/list", method = RequestMethod.GET)
-    public ModelAndView list(ModelAndView model, @RequestParam(name = "userId") Long userId) {
+    public ModelAndView list(ModelAndView model, @RequestParam(name = "userId") Integer userId) {
         Map<String, Object> queryMap = new HashMap<String, Object>();
         queryMap.put("offset", 0);
         queryMap.put("pageSize", 99999);
@@ -69,10 +69,10 @@ public class ExamPaperController {
     @RequestMapping(value = "/list", method = RequestMethod.POST)
     @ResponseBody
     public Map<String, Object> list(
-            @RequestParam(name = "examId", required = false) Long examId,
-            @RequestParam(name = "studentId", required = false) Long studentId,
+            @RequestParam(name = "examId", required = false) Integer examId,
+            @RequestParam(name = "studentId", required = false) Integer studentId,
             @RequestParam(name = "status", required = false) Integer status,
-            @RequestParam(name = "userId") Long userId,
+            @RequestParam(name = "userId") Integer userId,
             Page page
     ) {
         Map<String, Object> ret = new HashMap<String, Object>();
