+++ src/main/java/com/cg/dao/admin/MenuDao.java
@@ -58,7 +58,7 @@ public interface MenuDao {
      * @param id
      * @return
      */
-    int delete(Long id);
+    int delete(Integer id);
 
     /**
      * 获取某一分类的子菜单信息
@@ -66,7 +66,7 @@ public interface MenuDao {
      * @param parentId
      * @return
      */
-    List<Menu> findChildernList(Long parentId);
+    List<Menu> findChildernList(Integer parentId);
 
     /**
      * 根据菜单id获取菜单信息
