+++ src/main/java/com/cg/service/admin/ExamService.java
@@ -73,7 +73,7 @@ public interface ExamService {
      * @param id
      * @return
      */
-    Exam findById(Long id);
+    Exam findById(Integer id);
 
     /**
      * 考试统计数据更新操作
