+++ src/main/java/com/cg/service/admin/QuestionService.java
@@ -59,13 +59,13 @@ public interface QuestionService {
      * @param queryMap
      * @return
      */
-    int getQuestionNumByType(Map<String, Long> queryMap);
+    int getQuestionNumByType(Map<String, Integer> queryMap);
 
     /**
      * 试题信息按id查询
      * @param id
      * @return
      */
-    Question findById(Long id);
+    Question findById(Integer id);
 
 }
